requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.1
fastapi==0.122.0
h11==0.16.0
idna==3.11
ImageIO==2.37.2
lazy_loader==0.4
networkx==3.6
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==12.0.0
pydantic==2.12.4
pydantic_core==2.41.5
python-multipart==0.0.20
scikit-image==0.25.2
scipy==1.16.3
sniffio==1.3.1
starlette==0.50.0
tifffile==2025.10.16
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
